requirements.txt
fastapi
uvicorn
pytesseract
requests
Pillow
aiofiles
python-multipart
mangum           # optional

